use gurad clauses for rendering, remove hardcoding chain id, refactor naming and extra lines
